Move errors processing functions to utils #38
